+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog
 
 2.1.4 - ??/??/2018
 ==================
+- Added ssl_ciphers config option to only allow specific SSL ciphers
 - Fixed issue where total CPU and memory usage in processes were actually averages
 - Fixed error when trying to run check on a full interface node